Change "tapbranch queue" to "queue"
The priority queue in the huffman algorithm contains both tapleaves and
tapbranches, so naming it tapbranch queue is misleading. The priority
queue was previously described as a "node" queue, but it may be best to
drop the term node altogether. From the description of the huffman
algorithm, it is clear what element types are pushed to the queue.